+++ src/pix/client.rs
@@ -48,7 +48,7 @@ impl Client {
     }
 
     /// Write a pixel to the given stream.
-    pub fn write_pixel(&mut self, x: u32, y: u32, color: Color) -> Result<(), Error> {
+    pub fn write_pixel(&mut self, x: u16, y: u16, color: Color) -> Result<(), Error> {
         if self.binary {
             self.write_command(
                 &[
@@ -74,7 +74,7 @@ impl Client {
     }
 
     /// Read the size of the screen.
-    pub fn read_screen_size(&mut self) -> Result<(u32, u32), Error> {
+    pub fn read_screen_size(&mut self) -> Result<(u16, u16), Error> {
         // Read the screen size
         let data = self
             .write_read_command(b"SIZE")
@@ -87,10 +87,10 @@ impl Client {
         match re.captures(&data) {
             Some(matches) => Ok((
                 matches[1]
-                    .parse::<u32>()
+                    .parse::<u16>()
                     .expect("Failed to parse screen width, received malformed data"),
                 matches[2]
-                    .parse::<u32>()
+                    .parse::<u16>()
                     .expect("Failed to parse screen height, received malformed data"),
             )),
             None => Err(Error::new(
